pyproject.toml
[project]
name = "orak-2025-starter-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "ahk>=1.8.4",
    "ahk-binary>=2025.1.0",
    "aiohttp>=3.13.2",
    "anthropic>=0.72.0",
    "backoff>=2.2.1",
    "burnysc2>=5.0.5",
    "chardet>=5.2.0",
    "colorama>=0.4.6",
    "dacite>=1.9.2",
    "dataclass-wizard>=0.35.1",
    "diambra>=0.0.20",
    "diambra-arena>=2.2.7",
    "dill>=0.4.0",
    "dxcam>=0.0.5",
    "easyocr>=1.7.2",
    "fastmcp>=2.13.0.2",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.49.0",
    "gym==0.26.2",
    "gym-super-mario-bros>=7.4.0",
    "gymnasium==1.1.1",
    "imageio[ffmpeg]>=2.37.2",
    "langchain-chroma>=1.0.0",
    "langchain-openai>=1.0.2",
    "matplotlib>=3.10.7",
    "mcp[cli]>=1.20.0",
    "mss>=10.1.0",
    "multi-template-matching==1.6.6",
    "nes-py==8.2.1",
    "nest-asyncio>=1.6.0",
    "numpy==1.26.4",
    "omegaconf>=2.3.0",
    "openai>=2.7.1",
    "opencv-contrib-python>=4.11.0.86",
    "opencv-python>=4.11.0.86",
    "pillow>=10.3.0",
    "pyautogui>=0.9.54",
    "pyboy==2.5.2",
    "pydirectinput>=1.0.4 ; sys_platform == 'win32'",
    "pygame>=2.6.1",
    "pyobjc-framework-cocoa>=12.0 ; sys_platform == 'darwin'",
    "pyobjc-framework-quartz>=12.0 ; sys_platform == 'darwin'",
    "python-dotenv>=1.2.1",
    "pywin32>=311 ; sys_platform == 'win32'",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-image>=0.25.2",
    "segment-anything",
    "spacy>=3.8.7",
    "supervision>=0.26.1",
    "tenacity>=9.1.2",
    "termcolor>=3.2.0",
    "tiktoken>=0.12.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "ultralytics>=8.3.225",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git" }
